requirements.txt
# Web自动化测试
playwright==1.40.0
pytest-playwright==0.4.3

# 测试框架
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
pytest-timeout==2.2.0

# 测试报告
allure-pytest==2.13.2
allure-python-commons==2.13.2

# 数据驱动
pytest-datafiles==3.0.1

# 日志和工具
loguru==0.7.2
python-dotenv==1.0.0
pyyaml==6.0.1
openpyxl==3.1.2

# 其他工具
faker==20.1.0